refactor: Day12 리펙토링
