+++ client/src/components/HomePage/index.tsx
@@ -1,4 +1,7 @@
 import PostWriter from 'components/HomePage/PostWriter';
 import PostWriterModal from 'components/HomePage/PostWriterModal';
+import Post from 'components/HomePage/Post';
+import PostImageBox from 'components/HomePage/PostImageBox';
+import PostList from 'components/HomePage/PostList';
 
-export { PostWriter, PostWriterModal };
+export { PostWriter, PostWriterModal, Post, PostImageBox, PostList };
