+++ client/src/components/LoginPage/index.tsx
@@ -1,0 +1,5 @@
+import Footer from './Footer';
+import Greeter from './Greeter';
+import LoginBox from './LoginBox';
+
+export { Footer, Greeter, LoginBox };
