--- client/src/components/common/AlertSideBar.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const AlertSideBar = () => {
-  return null;
-};
-
-export default AlertSideBar;
