+++ client/src/components/common/SelectorSideBar.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { useRecoilValue } from 'recoil';
-import { rightModalStates, userData } from 'recoil/modal';
+import { rightModalStates, userData } from 'recoil/store';
 
 const SelectorSideBarContainer = styled.div<any>`
   width: inherit;
@@ -9,12 +9,14 @@ const SelectorSideBarContainer = styled.div<any>`
   box-shadow: -5px 2px 5px 0px rgb(0 0 0 / 24%);
 `;
 
-const SelectorSideBar: React.FC = () => {
+const SelectorSideBar = () => {
   const rightModalState = useRecoilValue(rightModalStates);
 
   if (rightModalState.rightModalFlag && rightModalState.selectorFlag) {
     return (
-      <SelectorSideBarContainer>This is SelectorSideBar</SelectorSideBarContainer>
+      <SelectorSideBarContainer>
+        This is SelectorSideBar
+      </SelectorSideBarContainer>
     );
   } else return null;
 };
