+++ client/src/components/common/index.tsx
@@ -1,0 +1,26 @@
+import AlarmSideBar from 'components/common/AlarmSideBar';
+import ChatSideBar from 'components/common/ChatSideBar';
+import SelectorSideBar from 'components/common/SelectorSideBar';
+import GroupSideBar from 'components/common/GroupSideBar';
+import InfoSideBar from 'components/common/InfoSideBar';
+import Gnb from 'components/common/Gnb';
+import SideBar from 'components/common/SideBar';
+import ProfilePhoto from 'components/common/ProfilePhoto';
+import { UserSearchBar, UserSearchModal } from 'components/common/UserSearch';
+import UserCard from 'components/common/UserCard';
+import InitUserData from 'components/common/InitUserData';
+
+export {
+  AlarmSideBar,
+  ChatSideBar,
+  SelectorSideBar,
+  GroupSideBar,
+  InfoSideBar,
+  Gnb,
+  SideBar,
+  ProfilePhoto,
+  UserSearchBar,
+  UserSearchModal,
+  UserCard,
+  InitUserData
+};
