--- client/src/components/index.tsx
@@ -1,26 +1,0 @@
-import AlarmSideBar from './common/AlarmSideBar';
-import ChatSideBar from './common/ChatSideBar';
-import SelectorSideBar from './common/SelectorSideBar';
-import GroupSideBar from './common/GroupSideBar';
-import InfoSideBar from './common/InfoSideBar';
-import Gnb from './common/Gnb';
-import SideBar from './common/SideBar';
-import ProfilePhoto from './common/ProfilePhoto';
-import { UserSearchBar, UserSearchModal } from './common/UserSearch';
-import UserCard from './common/UserCard';
-import Authority from './common/Authority';
-
-export {
-  AlarmSideBar,
-  ChatSideBar,
-  SelectorSideBar,
-  GroupSideBar,
-  InfoSideBar,
-  Gnb,
-  SideBar,
-  ProfilePhoto,
-  UserSearchBar,
-  UserSearchModal,
-  UserCard,
-  Authority
-};
