--- client/src/mock/index.tsx
@@ -1,52 +1,0 @@
-export const postMock = [
-  {
-    idx: 1,
-    useridx: 3,
-    secret: 0,
-    likenum: 30,
-    contents: '테스트1',
-    picture1: null,
-    picture2: null,
-    picture3: null
-  },
-  {
-    idx: 2,
-    useridx: 3,
-    secret: 0,
-    likenum: 40,
-    contents: '테스트2',
-    picture1: null,
-    picture2: null,
-    picture3: null
-  },
-  {
-    idx: 3,
-    useridx: 3,
-    secret: 0,
-    likenum: 50,
-    contents: '테스트3',
-    picture1: null,
-    picture2: null,
-    picture3: null
-  },
-  {
-    idx: 4,
-    useridx: 3,
-    secret: 0,
-    likenum: 0,
-    contents: '테스트4',
-    picture1: null,
-    picture2: null,
-    picture3: null
-  },
-  {
-    idx: 5,
-    useridx: 3,
-    secret: 0,
-    likenum: 1000,
-    contents: '테스트5',
-    picture1: null,
-    picture2: null,
-    picture3: null
-  }
-];
